pyproject.toml
[tool.poetry]
name = "awesome-fastapi-project"
version = "0.1.0"
description = ""
authors = ["Bohdan Agarkov <sidi0us.mail@gmail.com>"]
readme = "README.md"
packages = [{include = "awesome_fastapi_project"}]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.100.1"
uvicorn = "^0.23.2"
gunicorn = "^21.2.0"
tortoise-orm = "^0.19.3"
asyncpg = "^0.28.0"
aerich = "^0.7.2"
python-dotenv = "^1.0.0"
fastapi-utils = "^0.2.1"
httpx = "^0.24.1"
python-jose = "^3.3.0"
fastapi-jwt-auth = "^0.5.0"


[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"


[tool.aerich]
tortoise_orm = "awesome_fastapi_project.main.TORTOISE_ORM"
location = "./migrations"
src_folder = "./."
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
dev = "poetry:run_dev"
test = "poetry:run_pytest"